Alle controllers voor artikel in 1 package gegooid voor meer overzicht, update functie werkt, nu nog uitvogelen hoe je een object uit een lijst doorstuurd
