+++ Aplikassie2/src/main/java/com/rsvier/workshop/controllers/artikel/ArtikelGewijzigdController.java
@@ -1,4 +1,4 @@
-package com.rsvier.workshop.controllers;
+package com.rsvier.workshop.controllers.artikel;
 
 import com.rsvier.workshop.dao.ArtikelRepository;
 import com.rsvier.workshop.domein.Artikel;
@@ -10,16 +10,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/artikelgewijzigd")
-public class WijzigArtikelController {
+public class ArtikelGewijzigdController {
     CrudRepository repository;
 	
     @Autowired
-    public WijzigArtikelController(ArtikelRepository artikelRepository) {
+    public ArtikelGewijzigdController(ArtikelRepository artikelRepository) {
         repository = artikelRepository;
     }
 
     @PostMapping
-    public Artikel wijzigArtikel(Artikel artikel) {
+    public Artikel registerArtikel(Artikel artikel) {
+        artikel.setArtikelStatus(Artikel.ArtikelStatus.ACTIEF);
         repository.save(artikel);
         return artikel;
     }
