+++ Aplikassie2/src/main/java/com/rsvier/workshop/controllers/artikel/RegisterArtikelController.java
@@ -1,4 +1,4 @@
-package com.rsvier.workshop.controllers;
+package com.rsvier.workshop.controllers.artikel;
 
 import com.rsvier.workshop.dao.ArtikelRepository;
 import com.rsvier.workshop.domein.Artikel;
@@ -9,15 +9,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/artikeledit")
-public class ArtikelEditController {
+@RequestMapping("/artikeltoegevoegd")
+public class RegisterArtikelController {
     CrudRepository repository;
 	
     @Autowired
-    public ArtikelEditController(ArtikelRepository artikelRepository) {
+    public RegisterArtikelController(ArtikelRepository artikelRepository) {
         repository = artikelRepository;
     }
-    
+
     @PostMapping
     public Artikel registerArtikel(Artikel artikel) {
         artikel.setArtikelStatus(Artikel.ArtikelStatus.ACTIEF);
